In progress of making landing page. There are sizing issues that need to be addressed.
In the process of creating the landing page. The page currently has an image, some text with the title of the website and a brief description and a signup and a login in button. The issue is that this website breaks at different aspect ratios. I think that I am going to have to start again tomorrow and remove a lot of the css I have written and try and re do it.